Move guts of qgspointcloudrenderer.* to qgspointcloudlayerrenderer.*
