+++ python/core/auto_generated/pointcloud/qgspointcloudrenderer.sip.in
@@ -9,11 +9,6 @@
 
 
 
-
-
-
-
-
 /************************************************************************
  * This file has been generated automatically from                      *
  *                                                                      *
