+++ python/core/core_auto.sip
@@ -446,8 +446,9 @@
 %Include auto_generated/pointcloud/qgspointcloudattribute.sip
 %Include auto_generated/pointcloud/qgspointcloudattributemodel.sip
 %Include auto_generated/pointcloud/qgspointcloudlayer.sip
-%Include auto_generated/pointcloud/qgspointcloudrenderer.sip
+%Include auto_generated/pointcloud/qgspointcloudlayerrenderer.sip
 %Include auto_generated/pointcloud/qgspointclouddataprovider.sip
+%Include auto_generated/pointcloud/qgspointcloudrenderer.sip
 %Include auto_generated/metadata/qgsabstractmetadatabase.sip
 %Include auto_generated/metadata/qgslayermetadata.sip
 %Include auto_generated/metadata/qgslayermetadataformatter.sip
